Remove BigQueryServicesV3
This change merges the contents of the BigQueryServicesV3 interface
into the existing BigQueryServices interface and removes the v3
version of the interface.
