+++ sdks/java/io/google-cloud-platform/src/main/java/org/apache/beam/sdk/io/gcp/bigquery/BigQueryIO.java
@@ -436,7 +436,6 @@ public class BigQueryIO {
         .setValidate(true)
         .setWithTemplateCompatibility(false)
         .setBigQueryServices(new BigQueryServicesImpl())
-        .setBigQueryServicesV3(new BigQueryServicesV3Impl())
         .setParseFnV3(parseFn)
         .setApiVersion(3)
         .build();
@@ -620,7 +619,6 @@ public class BigQueryIO {
       abstract Builder<T> setUseLegacySql(Boolean useLegacySql);
       abstract Builder<T> setWithTemplateCompatibility(Boolean useTemplateCompatibility);
       abstract Builder<T> setBigQueryServices(BigQueryServices bigQueryServices);
-      abstract Builder<T> setBigQueryServicesV3(BigQueryServicesV3 bigQueryV3Services);
       // If value is set to 3, use BigQuery v3 API. Any other value will result in v2 read.
       abstract Builder<T> setApiVersion(Integer version);
       abstract Builder<T> setReadOptionsV3(ReadOptionsV3 options);
@@ -638,17 +636,12 @@ public class BigQueryIO {
     abstract boolean getValidate();
     @Nullable abstract Boolean getFlattenResults();
     @Nullable abstract Boolean getUseLegacySql();
-
     abstract Boolean getWithTemplateCompatibility();
-
     @Nullable abstract BigQueryServices getBigQueryServices();
-    @Nullable abstract BigQueryServicesV3 getBigQueryServicesV3();
     @Nullable abstract Integer getApiVersion();
     @Nullable abstract ReadOptionsV3 getReadOptionsV3();
-
     @Nullable abstract SerializableFunction<SchemaAndRecord, T> getParseFn();
     @Nullable abstract SerializableFunction<SchemaAndRowProto, T> getParseFnV3();
-
     @Nullable abstract Coder<T> getCoder();
 
     @VisibleForTesting
@@ -675,7 +668,7 @@ public class BigQueryIO {
         if (getApiVersion() == 3) {
           source = BigQueryV3TableSource.create(
               getTableProvider(), getParseFnV3(), coder, getBigQueryServices(),
-              getBigQueryServicesV3(), getReadOptionsV3());
+              getReadOptionsV3());
         } else {
           source = BigQueryTableSource.create(
               jobUuid,
@@ -1027,11 +1020,6 @@ public class BigQueryIO {
     @VisibleForTesting
     TypedRead<T> withTestServices(BigQueryServices testServices) {
       return toBuilder().setBigQueryServices(testServices).build();
-    }
-
-    @VisibleForTesting
-    TypedRead<T> withTestServicesV3(BigQueryServicesV3 testServices) {
-      return toBuilder().setBigQueryServicesV3(testServices).build();
     }
   }
